requirements.txt
flask
flask_login
flask_sqlalchemy
flask_restful
marshmallow